// packages/uikit/packages/Frameworks/Dialog/Dialog.types.ts
export interface ExitConfirmOptions {
  title?: string;
  caption?: string;
  confirmText?: string;
  cancelText?: string;
}

export interface ExitConfig {
  overlay?: boolean;
  escape?: boolean;
  confirm?: ExitConfirmOptions;
  icon?: boolean | ReactNode;
  footerButton?: boolean | ReactNode;
}

export interface OffsetConfig {
  x?: number | string;
  y?: number | string;
  width?: number | string;
}

export interface DialogHeaderConfig {
  icon?: IconProps;
  title?: ReactNode;
  caption?: ReactNode;
  content?: ReactNode;
}

export interface DialogFooterConfig {
  columnLayout?: boolean;
  content?: ReactNode;
  buttons?: (ButtonProps & { key?: string })[];
  buttonCondensed?: boolean | "pc";
}

export type PopoverHeaderConfig = DialogHeaderConfig;
export type PopoverFooterConfig = DialogFooterConfig;

export interface ModalHeaderConfig extends DialogHeaderConfig {
  centered?: boolean;
}
export type ModalFooterConfig = DialogFooterConfig;

export interface SheetHeaderConfig extends DialogHeaderConfig {
  iconReversed?: boolean;
}
export type SheetFooterConfig = DialogFooterConfig;

type DialogBaseConfig = ThemeSystemProps & BorderProps & RadiusProps;
// ...
export interface ModalConfig extends DialogBaseConfig {
  header?: ModalHeaderConfig;
  footer?: ModalFooterConfig;
  exit?: ExitConfig;
  custom?: ReactNode;
  defaultEnterAction?: () => void;
  maintainWhenRouting?: boolean;
  immersive?: boolean;
  outside?: DialogOutsideConfig;
  portalClassName?: string;
  portalStyle?: React.CSSProperties;
  contentClassName?: string;
  contentStyle?: React.CSSProperties;
  className?: string;
  style?: React.CSSProperties;
  popoverOwnerId?: string;
}
// ...
export interface DialogOutsideOptions {
  lockBodyScroll?: boolean;
  background?: ThemeBackgroundPaint;
  backgroundBlur?: BackgroundBlurValue;
  className?: string;
  style?: React.CSSProperties;
}

export type DialogOutsideConfig = boolean | DialogOutsideOptions;